+++ src/ft_max.c
@@ -6,11 +6,13 @@
 /*   By: qle-guen <qle-guen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/25 15:20:53 by qle-guen          #+#    #+#             */
-/*   Updated: 2016/04/05 23:42:17 by qle-guen         ###   ########.fr       */
+/*   Updated: 2016/04/05 23:53:07 by qle-guen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int			ft_max(int a, int b)
+#include <libft.h>
+
+size_t	ft_max(size_t a, size_t b)
 {
 	if (a > b)
 		return (a);
